+++ lib/lono/seed/base.rb
@@ -2,93 +2,116 @@ require "fileutils"
 require "memoist"
 require "yaml"
 
-# Subclasses must implement: setup, params, variables
+# Subclasses should implement:
+#
+#   variables - Returns String with content of varibles files.
+#   setup - Hook to do extra things like create IAM service roles.
+#   finish - Finish hook after config files have been created.
+#
+# Note there is no params method to hook. The Base class handles params well.
+#
 class Lono::Seed
   class Base
     include Lono::AwsServices
-    include Helpers
     include Lono::Conventions
     extend Memoist
 
     def initialize(blueprint, options)
       @blueprint, @options = blueprint, options
-
       @template, @param = template_param_convention(options)
-      @args = options[:args] || {} # hash
-      @args.symbolize_keys!
-
-      @written_files = []
-      puts "Seeding starter values for #{@blueprint.color(:green)} blueprint configs"
     end
 
     def run
+      check_dsl_type!
       setup
-      write_configs
+      create_params
+      create_variables
       finish
-      final_message
     end
 
-    # Should be implemented in subclass
-    def setup; end
-    def params; end
-    def variables; end
-
-    # Optionally implemented in subclasses
-    def params_form
-      :short # can be short, medium, or long
-    end
-
-    def finish; end
-
-  protected
-    def params_path
-      case params_form.to_sym
-      when :short
-        "configs/#{@blueprint}/params/#{Lono.env}.txt"
-      when :medium
-        "configs/#{@blueprint}/params/#{Lono.env}/#{@param}.txt"
-      else
-        "configs/#{@blueprint}/params/#{Lono.env}/#{@template}/#{@param}.txt"
+    # Always create params files
+    def create_params
+      with_each_template do |path|
+        create_param_file(path)
       end
     end
 
-    def variables_path
-      "configs/#{@blueprint}/variables/#{Lono.env}.rb"
+    def create_variables
+      return unless variables
+      dest_path = "configs/#{@blueprint}/variables/#{Lono.env}.rb"
+      write(dest_path, variables)
+      puts "Starter variables created: #{dest_path}"
     end
 
-    def base_params_path
-      params_path.sub("params/#{Lono.env}", "params/base")
+    def check_dsl_type!
+      # TODO: only support DSL type right now
     end
 
-    def base_variables_path
-      variables_path.sub("variables/#{Lono.env}", "variables/base")
+    def setup; end
+    def finish; end
+
+    # meant to be overriden by subclass
+    def variables
+      <<~EOL
+      # This is an empty starter variables file. Please refer to the blueprint's README for variables to set.
+      # Note some blueprints may not use variables.
+      EOL
     end
 
-    def write_configs
-      @params, @variables = params, variables # so it only gets called once
-      write_file(params_path, @params) if @params
-      write_file(variables_path, @variables) if @variables
+    def create_param_file(app_template_path)
+      parameters = parameters(app_template_path)
+
+      lines = []
+      lines << "# Required parameters:"
+      required(parameters).each do |name, data|
+        example = description_example(data["Description"])
+        lines << "#{name}=#{example}"
+      end
+      lines << "# Optional parameters:"
+      optional(parameters).each do |name, data|
+        lines << "# #{name}=#{data["Default"]}"
+      end
+      content = lines.join("\n") + "\n"
+
+      dest_path = "configs/#{@blueprint}/params/#{Lono.env}.txt" # only support environment level parameters for now
+      write(dest_path, content)
+      puts "Starter params created:    #{dest_path}"
+    end
+
+    def write(path, content)
+      FileUtils.mkdir_p(File.dirname(path))
+      IO.write(path, content)
+    end
+
+    def description_example(description)
+      default = '...'
+      return default unless description
+      md = description.match(/(Example|IE): (.*)/)
+      return default unless md
+      md[2]
+    end
+
+    def parameters(app_template_path)
+      builder = Lono::Template::Dsl::Builder.new(app_template_path, @blueprint, quiet: false)
+      template = builder.template
+      template["Parameters"]
+    end
+
+    def required(parameters)
+      parameters.reject { |logical_id, p| p["Default"] }
+    end
+
+    def optional(parameters)
+      parameters.select { |logical_id, p| p["Default"] }
     end
 
   private
-
-    def write_file(path, content)
-      FileUtils.mkdir_p(File.dirname(path))
-      IO.write(path, content)
-      @written_files << path
-    end
-
-    def final_message
-      config_list = @written_files.map { |i| "  * #{i}" }.join("\n")
-
-      puts <<~EOL
-        The #{@blueprint} blueprint configs are in:
-
-        #{config_list}
-
-        The starter values are specific to your AWS account. They meant to
-        be starter values. Please take a look, you may want to adjust the values.
-      EOL
+    def with_each_template
+      paths = Dir.glob("#{Lono.config.templates_path}/**/*.rb")
+      files = paths.select{ |e| File.file?(e) }
+      files.each do |path|
+        yield(path)
+      end
     end
   end
 end
